+++ app/telemetry/metrics.c
@@ -21,7 +21,7 @@
  *   fs.read, fs.write.
  */
 #include "../../oodar.h"
-#include "event.h"
+#include "../../core/event/event.h"
 #include <stdint.h>
 #include <string.h>
 #include <stdio.h>