+++ types.h
@@ -18,6 +18,5 @@ long long oo_monotonic_us(void);
 #include "types/types_res.h"
 #include "types/types_actor.h"
 #include "types/types_control.h"
-#include "types/types_q.h"
 
 #endif